# perf: replace nodejieba with @node-rs/jieba