--- a/lib/index.js
+++ b/lib/index.js
@@ -83,7 +83,7 @@ class NodePinyin extends Pinyin {
 
 function segment(hans) {
   try {
-    jieba = jieba || require("nodejieba");
+    jieba = jieba || require("@node-rs/jieba");
   } catch (ex) {
     console.error();
     console.error("    Segment need nodejieba, please run '$ npm install nodejieba'.");
@@ -93,7 +93,7 @@ function segment(hans) {
   }
   // 词语拼音库。
   PHRASES_DICT = PHRASES_DICT || require("../data/phrases-dict");
-  return jieba.cutSmall(hans, 4);
+  return jieba.cut(hans, 4);
 }
 
 const pinyin = new NodePinyin(PINYIN_DICT);