`@node-rs/jieba` is easy to install which no need `node-gyp, C++ toolchains` and without `postinstall scripts` which download binary files from Github release.

Also, better performnace.

### `@node-rs/jieba`

```
pinyin --segment 中文 x 359,323 ops/sec ±0.73% (98 runs sampled) memory usage: 136,351,744 bytes.
pinyin --segment --heteronym 中文 x 318,336 ops/sec ±1.78% (94 runs sampled) memory usage: 1,523,712 bytes.
pinyin --segment --heteronym --style NORMAL 中文 x 321,629 ops/sec ±1.28% (94 runs sampled) memory usage: 1,245,184 bytes.
```

### `nodejieba`
```
pinyin --segment 中文 x 340,823 ops/sec ±0.88% (95 runs sampled) memory usage: 249,540,608 bytes.
pinyin --segment --heteronym 中文 x 315,942 ops/sec ±1.14% (95 runs sampled) memory usage: 1,122,304 bytes.
pinyin --segment --heteronym --style NORMAL 中文 x 290,042 ops/sec ±2.11% (92 runs sampled) memory usage: 1,204,224 bytes.
```