@Brooooooklyn 2020-09-08
resolve https://github.com/hotoo/pinyin/issues/164