@hotoo 2021-04-11
这个性能上并没有很大的提升，而且也并不是纯 js 写的，而是需要依赖 rust，更换这个收益好像并不是很大。